# Optimize StationMap/Audience calculation