1000 * AscertainPlayerMarkets() (auf meinem PC):
release: 200ms
debug: 750ms

Erster Commit (AddMarket.Set()-Calls eingespart) :
release: 198ms
debug:   735ms


also noch nicht sonderlich viel ... Ersparnis (aber jeder "Method-Call" sorgt bei Debug ja fuer ... Debugframes ... da sind Einsparungen immer gut.