@GWRon 2022-01-21
Zweiter Commit (Struct replacement):

Vorher:
release
1000 * AscertainPlayerMarkets(): 198ms
100000 * GetTotalAntennaReceiverShare took: 610ms.
100000 * GetTotalCableNetworkReceiverShare took: 638ms.

debug: 
1000 * AscertainPlayerMarkets(): 735ms
100000 * GetTotalAntennaReceiverShare took: 2222ms.
100000 * GetTotalCableNetworkReceiverShare took: 2260ms.


Nun:
release
1000 * AscertainPlayerMarkets(): 178ms
100000 * GetTotalAntennaReceiverShare took: 560ms.
100000 * GetTotalCableNetworkReceiverShare took: 579ms.

debug: 
1000 * AscertainPlayerMarkets(): 691ms
100000 * GetTotalAntennaReceiverShare took: 2129ms.
100000 * GetTotalCableNetworkReceiverShare took: 2152ms.