@GWRon 2022-01-22
So ... ich hab noch ein wenig optimiert.

hier erstmal die "Rohergebnisse"
Original:
```
Release: 
Mit 1 Antenne:
1000 * AscertainPlayerMarkets took: 190ms.
500 * sections.GetAntennaShareMap()() took: 365ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 445ms.
100000 * GetTotalAntennaReceiverShare took: 559ms.
100000 * GetTotalSatelliteReceiverShare took: 23ms.
100000 * GetTotalCableNetworkReceiverShare took: 555ms.

Mit 18 Antennen:
1000 * AscertainPlayerMarkets took: 182ms.
500 * sections.GetAntennaShareMap()() took: 3069ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 3788ms.
100000 * GetTotalAntennaReceiverShare took: 650ms.
100000 * GetTotalSatelliteReceiverShare took: 26ms.
100000 * GetTotalCableNetworkReceiverShare took: 587ms.


Debug
Mit 1 Antenne:
1000 * AscertainPlayerMarkets took: 715ms.
500 * sections.GetAntennaShareMap()() took: 2451ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 2821ms.
100000 * GetTotalAntennaReceiverShare took: 2128ms.
100000 * GetTotalSatelliteReceiverShare took: 335ms.
100000 * GetTotalCableNetworkReceiverShare took: 2124ms.

Mit 18 Antennen:
1000 * AscertainPlayerMarkets took: 697ms.
500 * sections.GetAntennaShareMap()() took: 15391ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 20965ms.
100000 * GetTotalAntennaReceiverShare took: 2231ms.
100000 * GetTotalSatelliteReceiverShare took: 326ms.
100000 * GetTotalCableNetworkReceiverShare took: 2116ms.
```

Optimiert:
```
Release: 
Mit 1 Antenne:
1000 * AscertainPlayerMarkets took: 173ms.
500 * sections.GetAntennaShareMap()() took: 163ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 247ms.
100000 * GetTotalAntennaReceiverShare took: 623ms.
100000 * GetTotalSatelliteReceiverShare took: 7ms.
100000 * GetTotalCableNetworkReceiverShare took: 533ms.
[22:27:48] DBG      |    : App suspended.

Mit 18 Antennen:
1000 * AscertainPlayerMarkets took: 172ms.
500 * sections.GetAntennaShareMap()() took: 2592ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 3223ms.
100000 * GetTotalAntennaReceiverShare took: 680ms.
100000 * GetTotalSatelliteReceiverShare took: 6ms.
100000 * GetTotalCableNetworkReceiverShare took: 532ms.


Debug:
Mit 1 Antenne:
1000 * AscertainPlayerMarkets took: 709ms.
500 * sections.GetAntennaShareMap()() took: 1412ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 1821ms.
100000 * GetTotalAntennaReceiverShare took: 2161ms.
100000 * GetTotalSatelliteReceiverShare took: 328ms.
100000 * GetTotalCableNetworkReceiverShare took: 2108ms.

Mit 18 Antennen:
1000 * AscertainPlayerMarkets took: 756ms.
500 * sections.GetAntennaShareMap()() took: 10366ms.
500 * GetTotalAntennaReceiverShare() (invalidated data) took: 16367ms.
100000 * GetTotalAntennaReceiverShare took: 2199ms.
100000 * GetTotalSatelliteReceiverShare took: 327ms.
100000 * GetTotalCableNetworkReceiverShare took: 2112ms.
```

Wichtig ist hier eigentlich einmal "GetAntennaShareMap()" - weil das berechnet wird, wenn Daten invalidiert sind (hier koennte eventuell nur "partiell" invalidiert werden - also nur "angrenzende Bereiche"). Momentan sorgt die fuer "Mikroruckler" wenn sie ausgefuehrt wird und laenger dauert. Die Dauer ist abhaengig von der Anzahl der Satelliten.

Und ausserdem wichtig sind die "AscertainPlayerMarkets" weil die jeweils die "gecachten" Informationen abrufen - und bei den Broadcasts/Vorhersagen eingesetzt werden.

Deswegen stelle ich die hier nochmal gegenueber:
```
                                    1 Antenne                   18 Antennen
                                    Original    Optimiert       Original    Optimiert
Release
1000 * AscertainPlayerMarkets       190ms       173ms           182ms       172ms
500 * GetAntennaShareMap            365ms       163ms           3069ms      2592ms

Debug
1000 * AscertainPlayerMarkets       715ms       709ms           697ms       756ms
500 * GetAntennaShareMap            2451ms      1412            15391ms     10366ms
```
Die Zahlen sind natuerlich nicht wirklich gemittelt ... aber man sollte vorallem bei "GetAntennaShareMap()" nun schon einen ordentlichen Zeitgewinn erkennen koennen - gerade im "Debug" und Schnellvorlauf sollte da einiges an gesparten Millisekunden pro Testdurchgang herumkommen