@nittka 2022-01-23
Der letzte zusammengefasste Vergleich bezieht sich nur auf den entsprechenden Commit, nicht auf "gesamt vorher" und hinterher, oder? Kannst Du Dir den Zuwachs in Debug AscertainPlayermarkets bei 18 Antennen erklären?
Ansonsten spielt natürlich nicht nur die reine Berechnungszeit eine Rolle, sondern auch, ob eine Berechnung überhaupt nötig gewesen wäre.
Unabhängig davon: Danke für die Optimierungsarbeit.