@GWRon 2022-01-23
Der Zuwachs...wie gesagt mittel ich nicht mehrere Durchlaeufe und manchmal schwanken die Ausfuehrungszeiten...wenn gerade die Cpu ein wenig durch irgendwas "mitbelastet" wird.

Man kann es also nur als Tendenz sehen.


Und das "nur berechnen, was noetig"... Da bin ich gedanklich schon dran. Hier ist der einfachste Ansatz erstmal, nur zu Invalidieren, wenn Schnittpunkte mit existierenden bestehen, beim Entfernen nur genau berechnen, wenn Schnittpunkte bestehen, ansonsten einfach abziehen.
Und im weiteren Ansatz alle geschnittenenen Stationen und die hinzugefuegte/entfernte Antenne einzeln fuer eine genaue Berechnung heranzuziehen.


Ist im uebrigen ein sehr sehr aehnliches Problem zu einem, was ich gerade auf Arbeit angehen muss/kann.