@GWRon 2022-01-24
Ok, aus der "shareMap" ist nun ein "shareGrid" geworden ...damit sollte "konstanterer" Zugriff auf die Daten ermoeglicht werden - je voller aber die Karte, desto unvorteilhafter wird die alte Map - da pro Bildpunkt ein Objekt gespeichert wurde.

Neben den Objekten entfaellt nun auch der Zuwachs in der Lookup-Zeit mit steigender Anzahl an Sendemasten - dafuer aber bereits leicht erhoehte Lookupzeit, wenn nur eine Antenne genutzt wird - aber sollte marginal sein.

```
                                 | 1 Antenne                   | 18 Antennen                 |
                                 | Orig.   | Opt#1   | Opt#2   | Orig.   | Opt#1   | Opt#2   |
Release -------------------------------------------------------------------------------------|
1000 * AscertainPlayerMarkets    |   190ms |   173ms |   198ms |   182ms |   172ms |   192ms |
500 * GetAntennaShareMap         |   365ms |   163ms |    81ms |  3069ms |  2592ms |   186ms |
Debug ---------------------------------------------------------------------------------------|
1000 * AscertainPlayerMarkets    |   715ms |   709ms |  1087ms |   697ms |   756ms |   706ms |
500 * GetAntennaShareMap         |  2451ms |  1412ms |   787ms | 15391ms | 10366ms |  4494ms |

Original: Master-Stand
Opt#1: vorheriger Commit
Opt#2: Grid statt Map
```