@nittka 2022-01-25
>... - je voller die Karte, desto unvorteilhafter wurde die alte Map - da pro Bildpunkt ein Objekt gespeichert wurde.

Heißt das, bei voller Karte für alle Spieler explodiert die Ausführungszeit?